Refactor into ListUtils, support nested lists
