+++ storage/ListByteBuffer.cpp
@@ -3,6 +3,7 @@
 #include <algorithm>
 #include <type_traits>
 
+#include "ListView.h"
 #include "ListElementView.h"
 
 using namespace db;
@@ -104,4 +105,5 @@ template ListElementView ListByteBuffer<>::write(ListBufferTypeTag, const types:
 template ListElementView ListByteBuffer<>::write(ListBufferTypeTag, const types::Bool::Primitive&);
 template ListElementView ListByteBuffer<>::write(ListBufferTypeTag, const types::String::Primitive&);
 template ListElementView ListByteBuffer<>::write(ListBufferTypeTag, const types::Embedding::Primitive&);
+template ListElementView ListByteBuffer<>::write(ListBufferTypeTag, const ListView&);
 }
